Fix for mobile and exclude jvm
