+++ shared/build.gradle.kts
@@ -101,7 +101,6 @@ kotlin {
 
                 implementation(libs.permissions)
                 implementation(libs.permissions.notifications)
-                implementation(libs.permissions.compose)
 
                 api(libs.generativeai)
             }
@@ -116,6 +115,7 @@ kotlin {
             dependencies {
                 api(libs.firebase.mpp.auth)
                 api(libs.apollo.normalized.cache.sqlite)
+                implementation(libs.permissions.compose)
             }
         }
 
@@ -125,6 +125,7 @@ kotlin {
 
         androidMain {
             dependsOn(mobileMain)
+
             dependencies {
                 api(project(":proto"))
                 api(libs.androidx.lifecycle.viewmodel.ktx)
