+++ shared/src/wasmJsMain/kotlin/dev/johnoreilly/confetti/permissions/rememberNotificationPermissionState.wasmJs.kt
@@ -1,0 +1,8 @@
+package dev.johnoreilly.confetti.permissions
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun rememberNotificationPermissionState(notificationsActive: Boolean?): NotificationPermissionState {
+    return NotificationPermissionState.NotApplicable
+}
